fix(mobile): couple of more UI glitches on Android
