+++ apps/mobile/src/app/import-signers/private-key-success.tsx
@@ -1,30 +1,16 @@
-import { StyleSheet } from 'react-native'
-import { LinearGradient } from 'expo-linear-gradient'
 import { ImportSuccess } from '@/src/features/ImportPrivateKey/components/ImportSuccess'
 import React from 'react'
-import { useTheme, View } from 'tamagui'
+import { getTokenValue, View } from 'tamagui'
 import { useSafeAreaInsets } from 'react-native-safe-area-context'
+import { AbsoluteLinearGradient } from '@/src/components/LinearGradient'
 
 export default function ImportPrivateKeySuccess() {
-  const theme = useTheme()
-  const colors: [string, string] = [theme.success.get(), 'transparent']
   const { bottom } = useSafeAreaInsets()
   return (
-    <View style={{ flex: 1 }} paddingBottom={bottom}>
-      <LinearGradient colors={colors} style={styles.background} />
+    <View flex={1} paddingBottom={Math.max(bottom, getTokenValue('$4'))}>
+      <AbsoluteLinearGradient />
 
       <ImportSuccess />
     </View>
   )
 }
-
-const styles = StyleSheet.create({
-  background: {
-    position: 'absolute',
-    left: 0,
-    right: 0,
-    top: 0,
-    height: 300,
-    opacity: 0.1,
-  },
-})
