+++ apps/mobile/src/app/import-signers/private-key.tsx
@@ -1,13 +1,13 @@
 import React from 'react'
 import { ImportPrivateKey } from '@/src/features/ImportPrivateKey'
-import { View } from 'tamagui'
+import { getTokenValue, View } from 'tamagui'
 import { useSafeAreaInsets } from 'react-native-safe-area-context'
 
 function PrivateKeyImport() {
-  const insets = useSafeAreaInsets()
+  const { bottom } = useSafeAreaInsets()
 
   return (
-    <View paddingHorizontal={'$4'} flex={1} paddingBottom={insets.bottom}>
+    <View paddingHorizontal={'$4'} flex={1} paddingBottom={Math.max(bottom, getTokenValue('$4'))}>
       <ImportPrivateKey />
     </View>
   )
