+++ apps/mobile/src/components/LinearGradient/index.ts
@@ -1,0 +1,1 @@
+export { AbsoluteLinearGradient } from './LinearGradien'
